add exception handling

+++ app/src/main/java/pa/chan/shift_test/data/dto/CardInfoEntity.kt
@@ -27,8 +27,8 @@ data class CountryEntity(
     val name: String?,
     val emoji: String?,
     val currency: String?,
-    val latitude: Int?,
-    val longitude: Int?,
+    val latitude: Double?,
+    val longitude: Double?,
 )
 
 data class BankEntity(
