+++ app/src/main/java/pa/chan/shift_test/domain/model/CardInfoModel.kt
@@ -22,8 +22,8 @@ data class CountryModel(
     val name: String?,
     val emoji: String?,
     val currency: String?,
-    val latitude: Int?,
-    val longitude: Int?,
+    val latitude: Double?,
+    val longitude: Double?,
 )
 
 data class BankModel(
